Migrated test for store loader
